doen

+++ src/Ludo.Blazor/Models/PlayerSetup.cs
@@ -2,6 +2,6 @@ namespace Ludo.Blazor.Models;
 
 public class PlayerSetup
 {
-    public required int PlayerNr { get; set; }
+    public required byte PlayerNr { get; set; }
     public required bool CanRoll { get; set; }
 }
